[Rename, Refactor, Code] 디렉토리 구조 수정, 파일명 변경, 주석 삭제
