+++ components/trade/chart/marketDetail/MarketDetail.jsx
@@ -16,7 +16,7 @@ function SubIndicator({ label, value, valueStyle }) {
   );
 }
 
-export default function MarketDetailTable({ codeMap, ticker, numColor }) {
+export default function MarketDetail({ codeMap, ticker, numColor }) {
   return (
     <div className="h-[6.25rem] border-dashed border-[0.313rem] border-main bg-white box-border">
       <div className="flex flex-wrap ml-[0.125rem] gap-0.5 items-end">
