+++ components/vision/videos/VideosContainer.jsx
@@ -2,10 +2,10 @@ import { useQuery } from '@tanstack/react-query';
 import { Alert } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
 import axios from 'axios';
-import PendingSkeleton from '../PendingSkeleton';
+import PendingSkeleton from '../shared/PendingSkeleton';
 import VideoCards from './VideoCards';
 
-export default function Videos() {
+export default function VideosContainer() {
   const theme = useTheme();
   const {
     isPending,
