+++ layouts/Layout.jsx
@@ -1,10 +1,10 @@
-import NavBar from '@/components/layout/Navbar';
+import NavBarContainer from '@/components/layout/NavbarContainer';
 import React from 'react';
 
 export default function Layout({ children }) {
   return (
     <div>
-      <NavBar />
+      <NavBarContainer />
       <main>{children}</main>
     </div>
   );
