+++ pages/404.jsx
@@ -1,24 +1,13 @@
+import { DescriptionTypo } from '@/defaultTheme';
 import React from 'react';
-import { useRouter } from 'next/router';
-import { Box } from '@mui/material';
-import { DescriptionTypo, NGTypo } from '@/defaultTheme';
 
 export default function NotFound() {
-  const router = useRouter();
-
   return (
-    <Box
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'center',
-        alignItems: 'center',
-      }}
-    >
-      <DescriptionTypo>제작되지 않은 페이지입니다.</DescriptionTypo>
-      <button onClick={() => router.back()}>
-        <NGTypo>돌아가기</NGTypo>
-      </button>
-    </Box>
+    <div className="flex flex-col justify-center items-center">
+      <div className="opacity-0">block</div>
+      <DescriptionTypo>
+        올바르지 않은 경로입니다. 다른 페이지로 이동해주세요.
+      </DescriptionTypo>
+    </div>
   );
 }
